Fix projectile stuttering.
This commit simply disables the autoflush feature which is enabled by default in the window. The window is then flushed after the projectile has been moved.

An alternative fix would be to not create a new projectile circle every tick, and instead move the existing one but I found it to be a little more cumbersome relative to the existing codebase.